requirements-dev.txt
# Development Dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0
isort>=5.13.0

# Type stubs
types-PyYAML>=6.0.0
types-redis>=4.6.0
types-requests>=2.31.0
pandas-stubs>=2.1.0
